Extract Age() and Columns() to output package
